+++ lib/libesp32/Berry/generate/be_fixed_introspect.h
@@ -1,15 +1,16 @@
 #include "be_constobj.h"
 
 static be_define_const_map_slots(m_libintrospect_map) {
-    { be_const_key(members, -1), be_const_func(m_attrlist) },
-    { be_const_key(vcall, -1), be_const_func(m_vcall) },
     { be_const_key(get, -1), be_const_func(m_findmember) },
-    { be_const_key(set, 2), be_const_func(m_setmember) },
+    { be_const_key(toptr, 3), be_const_func(m_toptr) },
+    { be_const_key(set, -1), be_const_func(m_setmember) },
+    { be_const_key(fromptr, -1), be_const_func(m_fromptr) },
+    { be_const_key(members, 1), be_const_func(m_attrlist) },
 };
 
 static be_define_const_map(
     m_libintrospect_map,
-    4
+    5
 );
 
 static be_define_const_module(